fix: convert only if it's actually not a integrer with comparison operator
