+++ pkg/telemetrylogs/filter_expr_logs_test.go
@@ -430,7 +430,7 @@ func TestFilterExprLogs(t *testing.T) {
 			category:              "FREETEXT with conditions",
 			query:                 "critical NOT resolved status > open",
 			shouldPass:            true,
-			expectedQuery:         "WHERE (match(LOWER(body), LOWER(?)) AND NOT (match(LOWER(body), LOWER(?))) AND (attributes_number['status'] > ? AND mapContains(attributes_number, 'status') = ?))",
+			expectedQuery:         "WHERE (match(LOWER(body), LOWER(?)) AND NOT (match(LOWER(body), LOWER(?))) AND (toString(attributes_number['status']) > ? AND mapContains(attributes_number, 'status') = ?))",
 			expectedArgs:          []any{"critical", "resolved", "open", true},
 			expectedErrorContains: "",
 		},
